requirements.txt
pip freeze > requirements.txt

flask
gunicorn
python-dotenv
auth0-python
